+++ nyx/bytecode/MetaArea.h
@@ -8,7 +8,7 @@
 struct MetaArea {
 
     int bytecodeSize;
-    nyx::int8 *bytecodes;
+    nyx::int32 *bytecodes;
     std::vector<std::string> strings;
 };
